+++ Runtime/Model/BacktraceReport.cs
@@ -1,8 +1,7 @@
-﻿
-using Backtrace.Unity.Common;
+﻿using Backtrace.Unity.Common;
+using Backtrace.Unity.Extensions;
 using System;
 using System.Collections.Generic;
-using System.Linq;
 
 namespace Backtrace.Unity.Model
 {
@@ -106,6 +105,7 @@ namespace Backtrace.Unity.Model
             {
                 Message = exception.Message;
                 SetClassifier();
+                SetReportFingerPrintForEmptyStackTrace();
             }
             SetStacktraceInformation();
         }
@@ -138,6 +138,20 @@ namespace Backtrace.Unity.Model
             }
             MinidumpFile = minidumpPath;
             AttachmentPaths.Add(minidumpPath);
+        }
+
+        /// <summary>
+        /// Override default fingerprint for reports without faulting stack trace.
+        /// </summary>
+        internal void SetReportFingerPrintForEmptyStackTrace()
+        {
+            if (string.IsNullOrEmpty(Exception.StackTrace))
+            {
+                // set attributes instead of fingerprint to still allow our user to define customer
+                // fingerprints for reports without stack trace and apply deduplication rules in report flow.
+                Attributes["_mod_fingerprint"] = Exception.Message.OnlyLetters().GetSha();
+            }
+
         }
 
         internal BacktraceData ToBacktraceData(Dictionary<string, string> clientAttributes, int gameObjectDepth)