+++ Runtime/Model/BacktraceUnhandledException.cs
@@ -44,22 +44,43 @@ namespace Backtrace.Unity.Model
         /// Unhandled exception stack frames
         /// </summary>
         public List<BacktraceStackFrame> StackFrames = new List<BacktraceStackFrame>();
+        public BacktraceSourceCode SourceCode = null;
+
 
         public BacktraceUnhandledException(string message, string stacktrace) : base(message)
         {
             _message = message;
-
-            if (string.IsNullOrEmpty(stacktrace))
+            if (!string.IsNullOrEmpty(stacktrace))
             {
-                _stacktrace = new StackTrace(0, true).ToString();
+                _stacktrace = stacktrace;
+                ConvertStackFrames();
             }
             else
             {
-                _stacktrace = stacktrace.Trim();
-                ConvertStackFrames();
+                _stacktrace = string.Empty;
+                var backtraceStackTrace = new BacktraceStackTrace(null);
+                StackFrames = backtraceStackTrace.StackFrames;
             }
+            CreateUnhandledExceptionLogInformation();
             TrySetClassifier();
 
+        }
+
+        /// <summary>
+        /// Assign source code information to first stack frame of unhandled exception report
+        /// </summary>
+        private void CreateUnhandledExceptionLogInformation()
+        {
+            SourceCode = new BacktraceSourceCode()
+            {
+                Text = string.Format("Unity exception information\nMessage :{0}\nStack trace :{1}", _message, _stacktrace)
+            };
+            // assign log information to first stack frame
+            if (StackFrames.Count == 0)
+            {
+                return;
+            }
+            StackFrames.First().SourceCode = SourceCode.Id.ToString();
         }
 
         /// <summary>