+++ Tests/Runtime/DeduplicationTests.cs
@@ -1,5 +1,4 @@
-﻿using Backtrace.Unity;
-using Backtrace.Unity.Model;
+﻿using Backtrace.Unity.Model;
 using Backtrace.Unity.Types;
 using NUnit.Framework;
 using System;
@@ -8,6 +7,7 @@ using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
 using UnityEngine.TestTools;
+using Backtrace.Unity.Extensions;
 
 namespace Backtrace.Unity.Tests.Runtime
 {
@@ -103,6 +103,45 @@ namespace Backtrace.Unity.Tests.Runtime
             var sha2 = deduplicationStrategy2.GetSha();
 
             Assert.AreNotEqual(sha1, sha2);
+        }
+
+        [Test]
+        public void TestFingerprintBehavior_ShouldGenerateFingerprintForExceptionReportWithoutStackTrace_ShouldIncludeFingerprintInBacktraceReport()
+        {
+            // exception without stack trace might happened when exception occured because of
+            // invalid game object setting or via weird crash
+            // exception below has empty exception stack trace
+            var exception = new BacktraceUnhandledException("00:00:00 00/00/00 Unhandled exception", string.Empty);
+
+            var report = new BacktraceReport(exception);
+            Assert.AreEqual(exception.Message.OnlyLetters().GetSha(), report.Attributes["_mod_fingerprint"]);
+            var data = new BacktraceData(report, null);
+            Assert.IsNotEmpty(data.Attributes.Attributes["_mod_fingerprint"].ToString());
+        }
+
+
+        [Test]
+        public void TestFingerprintBehavior_ShouldGenerateFingerprintWithOnlyLetters_ShouldIncludeFingerprintInBacktraceReport()
+        {
+            var exception = new BacktraceUnhandledException("00:00:00 00/00/00 Unhandled exception", string.Empty);
+            var report = new BacktraceReport(exception);
+            Assert.AreEqual(exception.Message.OnlyLetters().GetSha(), report.Attributes["_mod_fingerprint"]);
+        }
+
+        [Test]
+        public void TestFingerprintBehavior_ShouldGenerateFingerprintAndShouldntRemoveLetters_ShouldIncludeFingerprintInBacktraceReport()
+        {
+            var exception = new BacktraceUnhandledException("Unhandled exception", string.Empty);
+            var report = new BacktraceReport(exception);
+            Assert.AreEqual(exception.Message.OnlyLetters().GetSha(), report.Attributes["_mod_fingerprint"]);
+        }
+
+        [Test]
+        public void TestFingerprintBehavior_ShouldntGenerateFingerprintForExistingStackTrace_ShouldIgnoreAttributeFingerprint()
+        {
+            var exception = new BacktraceUnhandledException("Unhandled exception", "foo()");
+            var report = new BacktraceReport(exception);
+            Assert.IsFalse(report.Attributes.ContainsKey("_mod_fingerprint"));
         }
     }
 }